Maps Locate Commit Working
